+++ src/sgs_ctx.c
@@ -1,6 +1,7 @@
 
 #include <stdarg.h>
 #include <ctype.h>
+#include <errno.h>
 
 #define SGS_INTERNAL
 #define SGS_REALLY_INTERNAL
@@ -570,7 +571,7 @@ static void dumpobj( SGS_CTX, sgs_VarObj* p )
 	buf[0] = 0;
 	while( *ci )
 	{
-		int osi = ((int)*ci) == (int)SOP_FLAGS ? ARRAY_SIZE( sgs_IfaceNames ) - 1 : ((int)*ci);
+		int osi = ((int)(size_t)*ci) == (int)(size_t)SOP_FLAGS ? ARRAY_SIZE( sgs_IfaceNames ) - 1 : ((int)(size_t)*ci);
 		if( *buf )
 			strcat( buf, "," );
 		strcat( buf, sgs_IfaceNames[ osi ] );
